+++ malachite-nz/src/bin_util/demo_and_bench/natural_polynomial/comparison/mod.rs
@@ -9,11 +9,7 @@
 use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
-    basic::register(runner);
-    comparison::register(runner);
-    conversion::register(runner);
+    cmp::register(runner);
 }
 
-mod basic;
-mod comparison;
-mod conversion;
+mod cmp;