+++ malachite-nz/src/test_util/bench/bucketers.rs
@@ -89,6 +89,29 @@ pub fn natural_polynomial_bit_bucketer(var_name: &str) -> Bucketer<'_, NaturalPo
     }
 }
 
+pub fn pair_natural_polynomial_max_bit_bucketer<'a>(
+    x_name: &'a str,
+    y_name: &'a str,
+) -> Bucketer<'a, (NaturalPolynomial, NaturalPolynomial)> {
+    Bucketer {
+        bucketing_function: &|(p, q)| {
+            usize::exact_from(max(
+                p.coefficients_asc()
+                    .iter()
+                    .map(SignificantBits::significant_bits)
+                    .sum::<u64>(),
+                q.coefficients_asc()
+                    .iter()
+                    .map(SignificantBits::significant_bits)
+                    .sum::<u64>(),
+            ))
+        },
+        bucketing_label: format!(
+            "max({x_name}'s total coefficient bits, {y_name}'s total coefficient bits)"
+        ),
+    }
+}
+
 pub fn integer_polynomial_bit_bucketer(var_name: &str) -> Bucketer<'_, IntegerPolynomial> {
     Bucketer {
         bucketing_function: &|p| {