+++ malachite-nz/tests/lib.rs
@@ -587,6 +587,9 @@ pub mod natural_polynomial {
         pub mod named;
         pub mod size;
     }
+    pub mod comparison {
+        pub mod cmp;
+    }
     pub mod conversion {
         pub mod from_coefficients_asc;
         pub mod from_natural;